ignite_chain/invoice/x/invoice/types/messages_invoice_test.go
package types

import (
	"testing"

	"invoice/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateInvoice_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateInvoice
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateInvoice{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateInvoice{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateInvoice_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateInvoice
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateInvoice{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateInvoice{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteInvoice_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteInvoice
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteInvoice{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteInvoice{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
